# Allow to call update_columns with strong params