### Summary

Resolves #39180 

Benchmark changes:
old = without the patch
new = with the patch
```
Warming up --------------------------------------
                 old   922.000  i/100ms
Calculating -------------------------------------
                 old      9.002k (± 4.2%) i/s -     45.178k in   5.027960s
Warming up --------------------------------------
                 new   933.000  i/100ms
Calculating -------------------------------------
                 new      9.160k (± 2.9%) i/s -     46.650k in   5.096912s

Comparison:
                 new:     9160.4 i/s
                 old:     9002.2 i/s - same-ish: difference falls within error
```

Script for benchmark:
```ruby
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  git_source(:github) { |repo| "https://github.com/#{repo}.git" }

  #gem "rails", github: "santib/rails", branch: "fix_39180_update_columns_2"
  gem "rails", github: "rails/rails"
  gem "benchmark-ips", "< 2.8"
  gem "byebug"
  gem "sqlite3"
end

require "active_record"
require "active_model"

ActiveRecord::Base.establish_connection adapter: "sqlite3", database: ":memory:"
@connection = ActiveRecord::Base.connection
@connection.create_table :posts, force: true do |t|
  t.string :title
end

class Post < ActiveRecord::Base
end

post = Post.create
i = 0

Benchmark.ips do |x|
  x.report("old") { post.update_columns(title: (i = i+1)) }

  x.report("new") { post.update_columns(title: (i = i+1)) }

  x.hold! 'temp_results'
  x.compare!
end
```